requirements.txt
streamlit
mysql-connector-python
twilio
email-validator
python-dotenv
reportlab
bcrypt
pdfkit
fpdf